+++ packages/studio/sdk/package.json
@@ -9,7 +9,9 @@
     "build": "yarn && yarn run -T tsc --build",
     "watch": "yarn && yarn run -T tsc --build --watch"
   },
-  "files": [
-    "dist"
-  ]
+  "devDependencies": {
+    "@types/express": "^4.17.13",
+    "@types/node": "^16.11.13",
+    "knex": "0.20.1"
+  }
 }
